+++ lib/features/invoice/helper/cart_invoice_mapper.dart
@@ -1,7 +1,8 @@
 import 'package:thepos/features/carts/data/models/cart.dart';
 import 'package:thepos/features/carts/data/models/cart_item.dart';
-import 'package:thepos/features/invoice/models/invoice.dart';
-import 'package:thepos/features/invoice/models/invoice_item.dart';
+
+import '../data/models/invoice.dart';
+import '../data/models/invoice_item.dart';
 
 mixin CartInvoiceMapper {
   static Invoice? createInvoiceFrom({required Cart cart}) {
