+++ test/invoice/remote_store_invoice_test.dart
@@ -1,9 +1,9 @@
 import 'dart:convert';
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:thepos/features/invoice/models/invoice.dart';
-import 'package:thepos/features/invoice/repositories/api_invoice/remote_store_invoice.dart';
-import 'package:thepos/features/invoice/repositories/api_invoice/remote_store_invoice_error.dart';
+import 'package:thepos/features/invoice/data/data_sources/api_invoice/remote_store_invoice.dart';
+import 'package:thepos/features/invoice/data/data_sources/api_invoice/remote_store_invoice_error.dart';
+import 'package:thepos/features/invoice/data/models/invoice.dart';
 
 import '../helpers/mock_client_stub.dart';
 import 'helpers/remote_store_invoice_sut.dart';
