+++ test/products/products_api/helpers/remote_loader_sut.dart
@@ -1,6 +1,6 @@
 import 'package:thepos/features/products/products_api/remote_products_loader.dart';
 
-import '../../helpers/mock_client_stub.dart';
+import '../../../helpers/mock_client_stub.dart';
 
 class RemoteLoaderSUT {
   RemoteLoaderSUT(this.client, this.loader);
