+++ test/products/products_api/remote_load_products_test.dart
@@ -4,7 +4,7 @@ import 'package:thepos/features/products/product.dart';
 import 'package:thepos/features/products/products_api/remote_products_loader.dart';
 import 'package:thepos/features/products/products_api/remote_products_loader_errors.dart';
 
-import '../helpers/mock_client_stub.dart';
+import '../../helpers/mock_client_stub.dart';
 import 'helpers/remote_loader_sut.dart';
 
 void main() {
